requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1 